Improve contracts (#31)
* update pragma

* add ethLint

* fix solidity lint

* add require messages

* change uint to uint256 and use return variable

* pragma to ^0.5.0

* uint to uin256

* fix doc

* allocate returns in memory

* add require messages

* fix tests

+++ contracts/Migrations.sol
@@ -1,9 +1,10 @@
 pragma solidity ^0.5.14;
 
 
+
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     modifier restricted() {
         if (msg.sender == owner) _;
@@ -13,7 +14,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
